requirements.txt
pynput
pyperclip
pillow
psutil
cryptography
